DropInfo object added to containers tag
DropInfo useful in custom drag adorner templates (selectors) for accessing visual tree. E.g.: in nested ListBox scenario, if dragging an item of inner ListBox onto outer ListBox, not only the selected item should be moved but the whole group. Therefore also the drag adorner should not only show the screen capture of the selected item but the screen capture of its parent, the ListBox.